% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaya.R
\name{emissions_factors}
\alias{emissions_factors}
\title{Get emission factors for different energy sources}
\usage{
emissions_factors()
}
\value{
a tibble of values for emissions factors, in million metric
        tons of carbon dioxide per quad of energy.
}
\description{
Get emission factors for different energy sources
}
