% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fuel_mix}
\alias{fuel_mix}
\title{Mix of fuels contributing to primary energy supply for many countries}
\format{A tibble containing 936 rows and 7 variables
\describe{
  \item{country}{Country name}
  \item{country_code}{Three-letter country code}
  \item{geography}{Geographic category: "nation", "region", or "world"}
  \item{year}{The year}
  \item{fuel}{The fuel: "Coal", "Natural Gas", "Oil", "Nuclear", "Hydro",
              and "Renewables"}
  \item{quads}{The number of quads of that fuel consumed in the given country
               or region and year}
  \item{pct}{The percentage of that country or region's total primary
             energy consumption from the fuel}
}}
\source{
\url{https://www.bp.com/en/global/corporate/energy-economics/statistical-review-of-world-energy/downloads.html}
}
\usage{
fuel_mix
}
\description{
A dataset containing the fuel mix of how many quads and what fraction of
total primary energy supply comes from coal, gas, oil, nuclear, and
renewable sources.
}
\keyword{datasets}
