% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaya.R
\name{get_fuel_mix}
\alias{get_fuel_mix}
\title{Get fuel mix for a country or region}
\usage{
get_fuel_mix(region_name, region_code = NULL, quiet = FALSE)
}
\arguments{
\item{region_name}{The name of a country or region to look up}

\item{region_code}{Optional three-letter country or region code to look up
instead of the `region_name`}

\item{quiet}{Suppress warnings if there is no data for that country or
region.}
}
\value{
a tibble of fuel mix for the country or region.
  That is, the number of quads of each fuel and the
  fraction of total primary energy coming from that fuel.
}
\description{
Get fuel mix for a country or region
}
