% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaya.R
\name{kaya_country_list}
\alias{kaya_country_list}
\title{Get a list of countries in the Kaya data}
\usage{
kaya_country_list()
}
\value{
a vector of country names
}
\description{
Get a list of countries in the Kaya data
}
