% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_kaya.R
\docType{data}
\name{kaya_data_files}
\alias{kaya_data_files}
\title{List of data files, to look up when loading data}
\format{An object of class \code{list} of length 3.}
\usage{
kaya_data_files
}
\description{
List of data files, to look up when loading data
}
\keyword{datasets}
