% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaya_plots.R
\name{kaya_plot}
\alias{kaya_plot}
\title{Plot kaya-identity variable}
\usage{
kaya_plot(kaya_data, variable, start_year = NULL, y_lab = NULL)
}
\arguments{
\item{kaya_data}{A tibble with kaya-identity data}

\item{variable}{The name of the variable to plot (character)}

\item{start_year}{The year to start highlighting the data (should correspond
to the beginning of the trend calculation)}

\item{y_lab}{Optional label for the y-axis}
}
\value{
A plot oblect.
}
\description{
Plot kaya-identity variable
}
