% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaya.R
\name{kaya_region_list}
\alias{kaya_region_list}
\title{Get a list of countries in the Kaya data}
\usage{
kaya_region_list()
}
\value{
a vector of country and region names
}
\description{
Get a list of countries in the Kaya data
}
\seealso{
regions
}
