% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kayadata.R
\docType{package}
\name{kayadata-package}
\alias{kayadata-package}
\title{kayadata package}
\description{
kayadata is a package for loading Kaya identity data for many countries
into R.
}
\section{License}{


 The \pkg{kayadata} package is open source licensed under the
 MIT License.
}

\section{Bug reports}{

\itemize{
 \item kayadata issue tracker (\url{https://github.com/jonathan-g/kayadata/issues})
}
}

