% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_kaya.R
\docType{data}
\name{kayadata.env}
\alias{kayadata.env}
\title{Environment for keeping package-level variables}
\format{An object of class \code{environment} of length 2.}
\usage{
kayadata.env
}
\description{
Environment for keeping package-level variables
}
\keyword{internal}
