% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_kaya.R
\name{load_energy_by_fuel}
\alias{load_energy_by_fuel}
\title{Load energy_by_fuel data and store it to kayadata:::energy_by_fuel}
\usage{
load_energy_by_fuel()
}
\description{
Load energy_by_fuel data and store it to kayadata:::energy_by_fuel
}
