% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_kaya.R
\name{load_kaya_if_necessary}
\alias{load_kaya_if_necessary}
\title{Load Kaya data and fuel mix data if necessary}
\usage{
load_kaya_if_necessary()
}
\description{
Load Kaya data and fuel mix data if necessary
}
