% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaya.R
\name{megawatts_per_quad}
\alias{megawatts_per_quad}
\title{The number of megawatts it takes to replace a quad.}
\usage{
megawatts_per_quad()
}
\value{
The number of megawatts equivalent to one quad per year.
}
\description{
The number of megawatts of average power output over a year to
produce one quad of energy
}
\examples{
mwe <- megawatts_per_quad()
mwe
}
