% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaya.R
\docType{data}
\name{mtoe}
\alias{mtoe}
\title{Conversion factor: quads per MTOE}
\format{An object of class \code{numeric} of length 1.}
\usage{
mtoe
}
\description{
Conversion factor: quads per MTOE
}
\keyword{internal}
