% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_kaya.R
\docType{data}
\name{nation_translations}
\alias{nation_translations}
\title{Translations to reconcile different ways of writing nation names in different
databases.}
\format{An object of class \code{list} of length 2.}
\usage{
nation_translations
}
\description{
Translations to reconcile different ways of writing nation names in different
databases.
}
\keyword{datasets}
