% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kaya.R
\name{plot_fuel_mix}
\alias{plot_fuel_mix}
\title{Plot fuel mix}
\usage{
plot_fuel_mix(fuel_mix)
}
\arguments{
\item{fuel_mix}{A tibble with the mixture of fuels:
\describe{
  \item{fuel}{The name of the fuel}
  \item{quads}{The number of quads per year the country consumes}
  \item{pct}{The percentage of the country's energy that comes from that fuel}
}}
}
\value{
A plot oblect.
}
\description{
Plot fuel mix
}
