% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaya.R
\name{project_top_down}
\alias{project_top_down}
\title{Get top-down projections of Kaya variables for one or more countries
or regions for a given year}
\usage{
project_top_down(region_name, year, quiet = FALSE, region_code = NULL)
}
\arguments{
\item{region_name}{The name of a country or region to look up}

\item{year}{The year to project to}

\item{quiet}{Suppress warnings if there is no data for that country or
region.}

\item{region_code}{Optional three-letter country or region code to look up
instead of the \code{region_name}}
}
\value{
a tibble of values for \emph{P}, \emph{G}, \emph{E}, \emph{F}, \emph{g}, \emph{e}, \emph{f}, and \emph{ef}
for each country or region:
\describe{
\item{region}{The name of the country or region}
\item{year}{The year}
\item{P}{Population, in billions}
\item{G}{Gross domestic product, in trillions of constant 2015 U.S.
dollars.}
\item{E}{Total primary energy consumption, in quads}
\item{F}{CO2 emissions from fossil fuel consumption, in millions of metric
tons }
\item{g}{Per-capita GDP, in thousands of constant 2015 U.S. dollars per
person.}
\item{e}{Energy intensity of the economy, in quads per trillion dollars.}
\item{f}{Emissions intensity of the energy supply, in million metric tons
per quad.}
\item{ef}{Emissions intensity of the economy, in metric tons per
million dollars of GDP.}
}
}
\description{
Get top-down projections of Kaya variables for one or more countries
or regions for a given year
}
\examples{
project_top_down("China", 2037)
project_top_down(region_code = "VNM", year = 2043)
}
\seealso{
\link{regions}
}
