% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kayadata.R
\name{regions}
\alias{regions}
\title{Aggregate regional data}
\description{
Problems with aggregate regional data
}
\details{
The World Bank is missing GDP data for a number of nations, such as Syria
and Taiwan.  Because of this and the incommensurability between the regions
used for aggregate statistics in the World Bank data and the BP data,
aggregate regional data (e.g., for the Middle East and Africa) should be
treated with caution.
This problem does not hold for individual nations, where missing data
appears as \code{NA} values.
}
