% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{td_trends}
\alias{td_trends}
\title{Top-down projections of trends in Kaya variables for many countries}
\format{A tibble containing 42 rows and 5 variables
\describe{
   \item{country}{Country name}
   \item{year}{The year}
  \item{P}{Trend in population, in percent per year}
  \item{G}{Trend in gross demestic product, in percent per year}
  \item{E}{Trend in total primary energy consumption, in percent per year}
  \item{F}{Trend in CO2 emissions, in percent per year}
}}
\source{
\url{https://www.eia.gov/outlooks/ieo/ieo_tables.php}
}
\usage{
td_trends
}
\description{
A dataset containing top-down projections of trends in P, G, and E,
from the EIA's International Energy Outlook 2017.
}
\keyword{datasets}
