% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{td_values}
\alias{td_values}
\title{Top-down projections of future Kaya variables for many countries}
\format{A tibble containing 336 rows and 5 variables
\describe{
   \item{country}{Country name}
   \item{year}{The year}
  \item{P}{Population, in billions}
  \item{G}{Gross demestic product, in trillions of constant 2010 U.S. dollars}
  \item{E}{Total primary energy consumption, in quads}
  \item{F}{Total CO2 emissions, in millions of metric tons of carbon}
}}
\source{
\url{https://www.eia.gov/outlooks/ieo/ieo_tables.php}
}
\usage{
td_values
}
\description{
A dataset containing top-down projections of P, G, and E, from the
EIA's International Energy Outlook 2017.
}
\keyword{datasets}
