% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaya.R
\name{top_down_trend}
\alias{top_down_trend}
\title{Get top-down trends for Kaya variables for a country}
\usage{
top_down_trend(country_name)
}
\arguments{
\item{country_name}{The name of a country to look up}
}
\value{
a tibble of trends for P, G, E, F, g, e, f, and ef for the country,
in percent per year.
}
\description{
Get top-down trends for Kaya variables for a country
}
