% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaya.R
\name{top_down_values}
\alias{top_down_values}
\title{Get top-down projections of Kaya variables for a country or region}
\usage{
top_down_values(region_name, region_code, quiet = FALSE)
}
\arguments{
\item{region_name}{The name of a country or region to look up}

\item{region_code}{Optional three-letter country or region code to look up
instead of the `region_name`}

\item{quiet}{Suppress warnings if there is no data for that country or
region.}
}
\value{
a tibble of values for P, G, E, F, g, e, f, and ef for the country
or region:
\describe{
  \item{region}{The name of the country or region}
  \item{P}{Population, in billions}
  \item{G}{Gross demestic product, in trillions of constant 2010 U.S.
           dollars.}
  \item{E}{Total primary energy consumption, in quads}
  \item{F}{CO2 emissions from fossil fuel consumption, in millions of metric
           tons }
  \item{g}{Per-capita GDP, in thousands of constant 2010 U.S. dollars per
           person.}
  \item{e}{Energy intensity of the economy, in quads per trillion dollars.}
  \item{f}{Emissions intensity of the energy supply, in million metric tons
           per quad.}
  \item{ef}{Emissions intensity of the economy, in metric tons per
            million dollars of GDP.}
}
}
\description{
Get top-down projections of Kaya variables for a country or region
}
