% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaya.R
\name{emissions_factors}
\alias{emissions_factors}
\title{Get emission factors for different energy sources}
\usage{
emissions_factors(collapse_renewables = TRUE)
}
\arguments{
\item{collapse_renewables}{Combine hydroelectricity and other renewables
into a single category.}
}
\value{
a tibble of values for emissions factors, in million metric
tons of carbon dioxide per quad of energy.
}
\description{
Get emission factors for different energy sources
}
\examples{
e_fac <- emissions_factors()
e_fac
}
