% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaya.R
\name{get_fuel_mix}
\alias{get_fuel_mix}
\title{Get fuel mix for a country}
\usage{
get_fuel_mix(country_name)
}
\arguments{
\item{country_name}{The name of a country to look up}
}
\value{
a tibble of fuel mix for the country.
  That is, the number of quads of each fuel and the
  fraction of total primary energy coming from that fuel.
}
\description{
Get fuel mix for a country
}
