% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaya.R
\name{get_kaya_data}
\alias{get_kaya_data}
\title{Get Kaya data for a country}
\usage{
get_kaya_data(country_name)
}
\arguments{
\item{country_name}{The name of a country to look up}
}
\value{
a tibble of Kaya identity data for the country
}
\description{
Get Kaya data for a country
}
