% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaya.R
\name{lookup_country_code}
\alias{lookup_country_code}
\title{Look up country name from country code}
\usage{
lookup_country_code(country_code, data = kayadata::kaya_data)
}
\arguments{
\item{country_code}{The three-letter country code}

\item{data}{Data frame in which to look up `country_code`}
}
\value{
The corresponding country name, or NULL if there is no such country
}
\description{
Look up country name from country code
}
\keyword{internal}
