% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaya.R
\name{lookup_region_code}
\alias{lookup_region_code}
\title{Look up country or region name from code}
\usage{
lookup_region_code(region_code, data = kayadata::kaya_data)
}
\arguments{
\item{region_code}{The three-letter country or region code}

\item{data}{Data frame in which to look up `region_code`}
}
\value{
The corresponding country or region name, or NULL if there is no
        such country or region
}
\description{
Look up country or region name from code
}
\keyword{internal}
