% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kaya.R
\name{plot_fuel_mix}
\alias{plot_fuel_mix}
\title{Plot fuel mix}
\usage{
plot_fuel_mix(fuel_mix, collapse_renewables = TRUE)
}
\arguments{
\item{fuel_mix}{A tibble with the mixture of fuels:
\describe{
  \item{fuel}{The name of the fuel}
  \item{quads}{The number of quads per year the country or region consumes}
  \item{frac}{The percentage of the country's energy that comes from that fuel}
}}

\item{collapse_renewables}{Combine Hydro and other Renewables into a single
category.}
}
\value{
A plot object.
}
\description{
Plot fuel mix
}
