% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kaya.R
\name{plot_kaya}
\alias{plot_kaya}
\title{Plot kaya-identity variable}
\usage{
plot_kaya(kaya_data, variable, start_year = NA, stop_year = NA,
  y_lab = NULL, log_scale = FALSE, trend_line = FALSE)
}
\arguments{
\item{kaya_data}{A tibble with kaya-identity data}

\item{variable}{The name of the variable to plot (character)}

\item{start_year}{The year to start highlighting the data (should correspond
to the beginning of the trend calculation)}

\item{stop_year}{The year to stop highlighting the data (should correspond
to the beginning of the trend calculation)}

\item{y_lab}{Optional label for the y-axis}

\item{log_scale}{Use log scale for y axis}

\item{trend_line}{Include a trend line}
}
\value{
A plot oblect.
}
\description{
Plot kaya-identity variable
}
