% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaya.R
\name{project_top_down}
\alias{project_top_down}
\title{Get top-down projections of Kaya variables for a country for a given year}
\usage{
project_top_down(country_name, year, country_code = NULL)
}
\arguments{
\item{country_name}{The name of a country to look up}

\item{year}{The year to project to}

\item{country_code}{Optional three-letter country code to look up instead
of the `country_name`}
}
\value{
a tibble of values for P, G, E, F, g, e, f, and ef for the country:
\describe{
  \item{country}{The name of the country}
  \item{year}{The year}
  \item{P}{Population, in billions}
  \item{G}{Gross demestic product, in trillions of constant 2010 U.S.
           dollars.}
  \item{E}{Total primary energy consumption, in quads}
  \item{F}{CO2 emissions from fossil fuel consumption, in millions of metric
           tons }
  \item{g}{Per-capita GDP, in thousands of constant 2010 U.S. dollars per
           person.}
  \item{e}{Energy intensity of the economy, in quads per trillion dollars.}
  \item{f}{Emissions intensity of the energy supply, in million metric tons
           per quad.}
  \item{ef}{Emissions intensity of the economy, in metric tons per
            million dollars of GDP.}
}
}
\description{
Get top-down projections of Kaya variables for a country for a given year
}
