% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{td_trends}
\alias{td_trends}
\title{Top-down projections of trends in Kaya variables for many countries and regions}
\format{A tibble containing 226 rows and 11 variables
\describe{
\item{region}{Country or region name}
\item{region_code}{Three-letter country or region code}
\item{geography}{Geographic category: "nation", "region", or "world"}
\item{P}{Trend in population, in fraction per year}
\item{G}{Trend in gross domestic product, in fraction per year}
\item{E}{Trend in total primary energy consumption, in fraction per year}
\item{F}{Trend in CO2 emissions, in fraction per year}
\item{g}{Trend in per-capita GDP, in fraction per year}
\item{e}{Trend in energy intensity of the economy, in fraction per year}
\item{f}{Trend in emissions intensity of the energy supply, in fraction per year}
\item{ef}{Trend in emissions intensity of the economy, in fraction per year}
}}
\source{
\url{https://www.eia.gov/outlooks/archive/ieo17/}
}
\usage{
td_trends
}
\description{
A dataset containing top-down projections of trends in P, G, and E,
from the EIA's International Energy Outlook 2017.
}
\keyword{datasets}
