% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaya.R
\name{top_down_trend}
\alias{top_down_trend}
\title{Get top-down trends for Kaya variables for a country or region, using
projections from U.S. Energy Information Administration's International
Energy Outlook report.}
\usage{
top_down_trend(region_name, region_code = NULL)
}
\arguments{
\item{region_name}{The name of a country or region to look up}

\item{region_code}{Optional three-letter country or region code to look up
instead of the `region_name`}
}
\value{
a tibble of trends for P, G, E, F, g, e, f, and ef for the country,
or region in percent per year.
}
\description{
Get top-down trends for Kaya variables for a country or region, using
projections from U.S. Energy Information Administration's International
Energy Outlook report.
}
